requirements.txt
pandas
numpy
joblib
scikit-learn
tensorflow==2.16.1
matplotlib
streamlit
plotly
python-dotenv
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq
faiss-cpu
ragas
datasets
sentence-transformers
pypdf
docx2txt
fastapi
uvicorn
sqlalchemy